requirements.txt
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
docutils==0.20.1
enchant==0.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.3.0
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.1.0
nh3==0.2.14
pkginfo==1.9.6
pycparser==2.21
pygments==2.17.2
python-stdnum==1.20
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
SecretStorage==3.3.3
tqdm==4.67.1
twine==4.0.2
typing-extensions==4.8.0
urllib3==2.1.0
work-set-clustering==0.1.0
zipp==3.17.0
